test(AccountServiceTest): refactor tests to support BigDecimal balance type
- Updated AccountServiceTest to handle Account balance as BigDecimal instead of float.
- Refactored test setup and assertions to work with BigDecimal values for balance.
- Ensured that the tests for account retrieval, creation, and exception handling are consistent with the new BigDecimal type for balance.

+++ src/test/java/com/ng_billing/bank_management/service/AccountServiceTest.java
@@ -12,6 +12,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -30,7 +31,7 @@ class AccountServiceTest {
 
     @BeforeEach
     void setUp() {
-        account = new Account(234, 170.07f);
+        account = new Account(234, BigDecimal.valueOf(170.07f));
     }
 
     @Test
